Implement weather alert API

+++ src/Triggers/Conditions/WeatherTriggerCondition.cs
@@ -2,11 +2,11 @@
 {
     public sealed class WeatherTriggerCondition
     {
-        public double Clouds { get; }
-        public double Humidity { get; }
-        public double Pressure { get; }
-        public double Temperature { get; }
-        public double WindDirection { get; }
-        public double WindSpeed { get; }
+        public float Clouds { get; }
+        public float Humidity { get; }
+        public float Pressure { get; }
+        public float Temperature { get; }
+        public float WindDirection { get; }
+        public float WindSpeed { get; }
     }
 }
